Update phpthumbsup.class.php
This update allows you to use another (sub)domain for the images requests. This is useful when it comes to loading speed on a website with several requests. You can setup the system setting "base_url" with a full URL.
The only required thing is, after the domain; you should keep something like "phpthumbsup/" as fictive part of the URL. 
Also besides that, the (sub)domain should point to your MODX installation too. Because it needs MODX to run the code.